+++ src/pages/SchoolCreate/useSchoolCreate.ts
@@ -4,13 +4,11 @@ import { School, SchoolParams } from "@/types/school";
 import { Tag } from "@/types/tag";
 import { ref, onMounted } from "vue";
 import { createAxiosInstance } from "@/utils/axiosinstance";
-import { useMeStore } from "@/store/me";
 
 export function useSchoolCreate() {
     const alertStore = useAlertStore();
     const router = useRouter();
     const axiosInstance = createAxiosInstance();
-    const meStore = useMeStore();
 
     const pages = [
         { name: "スクール一覧", href: "/schools", current: false },
@@ -46,16 +44,6 @@ export function useSchoolCreate() {
         }
     };
 
-    const fetchMe = async () => {
-        try {
-            const { data } = await axiosInstance.get("/me");
-            meStore.setMe(data);
-        } catch (err) {
-            window.location.href = "/error";
-            return;
-        }
-    };
-
     const fetchTagOptions = async () => {
         try {
             const { data } = await axiosInstance.get("tag");
@@ -66,7 +54,6 @@ export function useSchoolCreate() {
     };
 
     onMounted(async () => {
-        await fetchMe();
         await fetchTagOptions();
     });
     return { pages, school, tagOptions, createSchool };
