+++ src/pages/Schools/useSchools.ts
@@ -3,14 +3,12 @@ import { School } from "@/types/school";
 import { ref, onMounted } from "vue";
 import { useAlertStore } from "@/store/alert";
 import { createAxiosInstance } from "@/utils/axiosinstance";
-import { useMeStore } from "@/store/me";
 
 export function useSchools() {
     const alertStore = useAlertStore();
     const router = useRouter();
     const pages = [{ name: "スクール一覧", href: "/schools", current: true }];
     const axiosInstance = createAxiosInstance();
-    const meStore = useMeStore();
 
     const schools = ref<School[]>([]);
 
@@ -23,18 +21,7 @@ export function useSchools() {
         }
     };
 
-    const fetchMe = async () => {
-        try {
-            const { data } = await axiosInstance.get("/me");
-            meStore.setMe(data);
-        } catch (err) {
-            window.location.href = "/error";
-            return;
-        }
-    };
-
     onMounted(async () => {
-        await fetchMe();
         await fetchSchools();
     });
