+++ src/pages/UserCreate/useUserCreate.ts
@@ -1,15 +1,12 @@
-import { onMounted } from "vue";
 import { useAlertStore } from "@/store/alert";
 import { useRouter } from "vue-router";
 import { User, userRole } from "@/types/user";
 import { createAxiosInstance } from "@/utils/axiosinstance";
-import { useMeStore } from "@/store/me";
 
 export function useUserCreate() {
     const alertStore = useAlertStore();
     const router = useRouter();
     const axiosInstance = createAxiosInstance();
-    const meStore = useMeStore();
 
     const pages = [
         { name: "ユーザー一覧", href: "/users", current: false },
@@ -23,16 +20,6 @@ export function useUserCreate() {
         role: userRole.Normal,
     };
 
-    const fetchMe = async () => {
-        try {
-            const { data } = await axiosInstance.get("/me");
-            meStore.setMe(data);
-        } catch (err) {
-            window.location.href = "/error";
-            return;
-        }
-    };
-
     const createUser = async (params: User) => {
         try {
             await axiosInstance.post("/user", params);
@@ -42,10 +29,6 @@ export function useUserCreate() {
             alertStore.showErrorAlert();
         }
     };
-
-    onMounted(async () => {
-        await fetchMe();
-    });
 
     return { pages, user, createUser };
 }
