+++ src/router/index.ts
@@ -17,6 +17,11 @@ import TagDetail from "../pages/TagDetail/TagDetail.vue";
 import Signup from "../pages/Signup/Signup.vue";
 
 import Error from "../pages/Error/Error.vue";
+import { createAxiosInstance } from "@/utils/axiosinstance";
+import { useMeStore } from "@/store/me";
+import { userRole } from "@/types/user";
+
+const axiosInstance = createAxiosInstance();
 
 const routes: RouteRecordRaw[] = [
     {
@@ -43,56 +48,100 @@ const routes: RouteRecordRaw[] = [
         path: "/counselings",
         name: "Counselings",
         component: Counselings,
+        meta: {
+            requiresAuth: true,
+            allowedRoles: [userRole.Admin, userRole.Staff],
+        },
     },
     {
         path: "/userCreate",
         name: "UserCreate",
         component: UserCreate,
+        meta: {
+            requiresAuth: true,
+            allowedRoles: [userRole.Admin],
+        },
     },
     {
         path: "/users",
         name: "Users",
         component: Users,
+        meta: {
+            requiresAuth: true,
+            allowedRoles: [userRole.Admin],
+        },
     },
     {
         path: "/users/:userId",
         name: "UserDetail",
         component: UserDetail,
+        meta: {
+            requiresAuth: true,
+            allowedRoles: [userRole.Admin],
+        },
     },
     {
         path: "/schools",
         name: "Schools",
         component: Schools,
+        meta: {
+            requiresAuth: true,
+            allowedRoles: [userRole.Admin],
+        },
     },
     {
         path: "/schools/:schoolId",
         name: "SchoolDetail",
         component: SchoolDetail,
+        meta: {
+            requiresAuth: true,
+            allowedRoles: [userRole.Admin],
+        },
     },
     {
         path: "/schoolCreate",
         name: "SchoolCreate",
         component: SchoolCreate,
+        meta: {
+            requiresAuth: true,
+            allowedRoles: [userRole.Admin],
+        },
     },
     {
         path: "/schools/:schoolId/edit",
         name: "SchoolEdit",
         component: SchoolEdit,
+        meta: {
+            requiresAuth: true,
+            allowedRoles: [userRole.Admin],
+        },
     },
     {
         path: "/tags",
         name: "Tags",
         component: Tags,
+        meta: {
+            requiresAuth: true,
+            allowedRoles: [userRole.Admin],
+        },
     },
     {
         path: "/tagCreate",
         name: "TagCreate",
         component: TagCreate,
+        meta: {
+            requiresAuth: true,
+            allowedRoles: [userRole.Admin],
+        },
     },
     {
         path: "/tags/:tagId",
         name: "TagDetail",
         component: TagDetail,
+        meta: {
+            requiresAuth: true,
+            allowedRoles: [userRole.Admin],
+        },
     },
     {
         path: "/error",
@@ -109,6 +158,45 @@ const routes: RouteRecordRaw[] = [
 const router = createRouter({
     history: createWebHistory(),
     routes,
+});
+
+const fetchMe = async () => {
+    const meStore = useMeStore();
+    try {
+        const { data } = await axiosInstance.get("/me");
+        meStore.setMe(data);
+    } catch (err) {
+        window.location.href = "/error";
+    }
+};
+
+router.beforeEach(async (to, from, next) => {
+    const meStore = useMeStore();
+
+    if (to.matched.some((record) => record.meta.requiresAuth)) {
+        await fetchMe();
+
+        // 認証チェック
+        if (!meStore.me.isAuthenticated) {
+            window.location.href = "/error";
+            return;
+        }
+
+        // 権限チェック
+        if (to.meta.allowedRoles) {
+            if (
+                Array.isArray(to.meta.allowedRoles) &&
+                !to.meta.allowedRoles.includes(meStore.me.role)
+            ) {
+                window.location.href = "/error";
+                return;
+            }
+        }
+
+        next();
+    } else {
+        next();
+    }
 });
 
 router.afterEach((to) => {
